requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
redis==5.0.1
faiss-cpu==1.8.0
numpy
sentence-transformers==2.2.2
torch
torchvision
groq
streamlit
